Move machine utils into main hierarchy
